Fix x86 optimizers, demo-tests: ``` :: After:  | | Entry point VIP:       0x1337  | | Stack pointer:         -0x8  | | Already visited?:      N  | | ------------------------  | | 0000: [ PSEUDO ]     -0x8     movd     eax          0x1  | | 0001: [ PSEUDO ]     -0x8     strd     $sp          -0x4         0x0  | | 0002: [ PSEUDO ]     -0x8     strd     $sp          -0x8         0x1  | | 0003: [ PSEUDO ]     -0x8     vpinrd   eax  | | 0004: [ PSEUDO ]     -0x8     vexitq   0x0 ```

+++ VTIL-Architecture/routine/basic_block.hpp
@@ -259,7 +259,7 @@ namespace vtil
 		// or the beginning of the basic block and the index of the stack instance.
 		//
 		uint32_t sp_index = 0;
-		int64_t sp_offset = 0;
+		intptr_t sp_offset = 0;
 
 		// Last temporary index used.
 		//
